use_natural_primary_keys must imply use_natural_foreign_keys
When dumping data, if we have use_natural_primary_keys, the
pk will be omitted from the serialized representation wherever
applicable.  Therefore, the pks's values will be chosen anew
when reloading the data.

However, if we also have use_natural_foreign_keys==False, foreign
references will be serialized using the current pk's value; and
therefore deserialized using it as well.  This old value may not
be the same as the new value.