serialized natural keys in multi-table inheritance models (revised)
when serializing with natural primary keys, it is important to treat
specially the case of a pk which is a relation with parent_link=True.
This most frequently happens with multi-table inheritance, but there
are other possibilities as well.  In that case, the field used as pk
must be serialized even if it was marked serialize=False.